+++ frontend/src/app/app.routes.ts
@@ -6,28 +6,26 @@ export const routes: Routes = [
   {
     path: ROUTES.landing,
     loadComponent: () =>
-      import('./features/landing/landing.component').then(
-        m => m.LandingComponent
-      ),
+      import('./pages/landing/landing.component').then(m => m.LandingComponent),
   },
   {
     path: ROUTES.paymentSelection,
     loadComponent: () =>
-      import('./features/payment-selection/payment-selection.component').then(
+      import('./pages/payment-selection/payment-selection.component').then(
         m => m.PaymentSelectionComponent
       ),
   },
   {
     path: ROUTES.orderTracking,
     loadComponent: () =>
-      import('./features/order-tracking/order-tracking.component').then(
+      import('./pages/order-tracking/order-tracking.component').then(
         m => m.OrderTrackingComponent
       ),
   },
   {
     path: ROUTES.splitPayment,
     loadComponent: () =>
-      import('./features/split-payment/split-payment.component').then(
+      import('./pages/split-payment/split-payment.component').then(
         m => m.SplitPaymentComponent
       ),
   },
@@ -35,12 +33,12 @@ export const routes: Routes = [
     path: ROUTES.orderTrackingQRcode,
     loadComponent: () =>
       import(
-        './features/order-tracking-qrcode/order-tracking-qrcode.component'
+        './pages/order-tracking-qrcode/order-tracking-qrcode.component'
       ).then(m => m.OrderTrackingQRcodeComponent),
   },
   {
     path: ROUTES.menu,
     loadComponent: () =>
-      import('./features/menu/menu.component').then(m => m.MenuComponent),
+      import('./pages/menu/menu.component').then(m => m.MenuComponent),
   },
 ];
