+++ frontend/src/app/core/models/basket.model.ts
@@ -1,0 +1,6 @@
+import { BasketItem } from './item.model';
+
+export type Basket = {
+  _id?: string;
+  items: BasketItem[];
+};
