+++ frontend/src/app/core/models/interfaces/category.ts
