--- frontend/src/app/features/landing/landing.component.html
@@ -1,1 +1,0 @@
-<app-option-selection-layout [options]="options">
