--- frontend/src/app/features/menu/menu.component.scss
@@ -1,68 +1,0 @@
-:host {
-  width: 100%;
-  height: 100%;
-
-  display: grid;
-  grid-template: 1fr 10vh / 22vw 1fr;
-}
-
-aside {
-  background-color: var(--secondary-color);
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 16px;
-}
-
-#logo-img {
-  width: auto;
-  height: 18vh;
-}
-
-#categories-container {
-  width: 100%;
-  height: 100%;
-
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 2vmin;
-  padding: 2vmin;
-}
-
-main {
-  text-align: center;
-}
-
-#filter-button {
-  align-self: start;
-}
-
-#items-container {
-  height: 100%;
-
-  border-left: black 1px solid;
-
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-auto-rows: max-content;
-  justify-items: center;
-  gap: 2vmin;
-  padding: 2vmin;
-}
-
-footer {
-  grid-column: 1 / -1;
-
-  background-color: var(--secondary-color);
-  border-top: black 1px solid;
-
-  box-shadow: var(--base-shadow);
-  color: var(--primary-color);
-
-  display: flex;
-  justify-content: end;
-  align-items: center;
-  padding: 4vmin;
-  gap: 4vmin;
-}
