--- frontend/src/app/features/order-tracking/order-tracking.component.html
@@ -1,8 +1,0 @@
-<div class="order-tracking">
-    <img src="logo.webp" alt="logo" class="logo">
-    <div class="content">
-        <fa-icon [icon]="currentIcon" class="icon"></fa-icon>
-        <app-progress-bar [stepNumber]="stepNumber" [currentStep]="currentStep"></app-progress-bar>
-        <h2 class="status">{{status}}</h2>
-    </div>
-</div>
