--- frontend/src/app/features/order-tracking/order-tracking.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { OrderTrackingComponent } from './order-tracking.component';
-
-describe('OrderTrackingComponent', () => {
-  let component: OrderTrackingComponent;
-  let fixture: ComponentFixture<OrderTrackingComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [OrderTrackingComponent]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(OrderTrackingComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
