--- frontend/src/app/features/payment-selection/payment-selection.component.html
@@ -1,1 +1,0 @@
-<app-option-selection-layout [options]="options">
