+++ frontend/src/app/pages/landing/landing.component.html
@@ -1,0 +1,1 @@
+<app-option-selection-layout [options]="options"></app-option-selection-layout>
