+++ frontend/src/app/pages/menu/components/footer/footer.component.scss
@@ -1,0 +1,16 @@
+footer {
+  width: 100%;
+  height: var(--footer-height);
+  background-color: var(--secondary-color);
+
+  box-shadow: var(--base-shadow);
+  border-top: var(--base-border);
+
+  color: var(--primary-color);
+
+  display: flex;
+  justify-content: end;
+  align-items: center;
+  padding: 4vmin;
+  gap: 4vmin;
+}
