+++ frontend/src/app/pages/menu/menu.component.scss
@@ -1,0 +1,60 @@
+main {
+  width: 100%;
+  height: calc(100% - var(--footer-height));
+
+  display: grid;
+  grid-template: 1fr / 22vw 1fr;
+}
+
+aside {
+  background-color: var(--secondary-color);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 16px;
+  box-shadow: var(--base-shadow);
+  border-right: var(--base-border);
+}
+
+#logo-img {
+  width: auto;
+  height: 18vh;
+}
+
+#categories-container {
+  width: 100%;
+  height: 100%;
+
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 2vmin;
+  padding: 2vmin;
+}
+
+#filter-button {
+  align-self: start;
+}
+
+#items-container {
+  height: 80%;
+
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  grid-auto-rows: max-content;
+  justify-items: center;
+  gap: 2vmin;
+}
+
+.menu {
+  padding: 2vmin;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+
+  &-header {
+    h2 {
+      margin: 1rem 0;
+    }
+  }
+}
