+++ frontend/src/app/pages/order-tracking-qrcode/order-tracking-qrcode.component.scss
