+++ frontend/src/app/pages/order-tracking/order-tracking.component.scss
@@ -1,0 +1,52 @@
+.order-tracking {
+  position: relative;
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: var(--secondary-color);
+}
+
+.logo {
+  position: absolute;
+  top: 20px;
+  left: 20px;
+  width: 100px;
+}
+
+.content {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 2rem;
+  padding: 4rem;
+
+  .icon {
+    font-size: 10rem;
+    color: var(--accent-color);
+    animation: bounce 1s infinite;
+  }
+
+  app-progress-bar {
+    width: 100%;
+  }
+
+  .status {
+    font-size: 2rem;
+    text-align: center;
+    color: var(--text-color);
+  }
+}
+
+@keyframes bounce {
+  0%,
+  100% {
+    transform: translateY(0);
+  }
+  50% {
+    transform: translateY(-20px);
+  }
+}
