+++ frontend/src/app/pages/payment-selection/payment-selection.component.html
@@ -1,0 +1,1 @@
+<app-option-selection-layout [options]="options"></app-option-selection-layout>
