+++ frontend/src/app/pages/payment-selection/payment-selection.component.spec.ts
@@ -1,0 +1,22 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { PaymentSelectionComponent } from './payment-selection.component';
+
+describe('PaymentSelectionComponent', () => {
+  let component: PaymentSelectionComponent;
+  let fixture: ComponentFixture<PaymentSelectionComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [PaymentSelectionComponent],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(PaymentSelectionComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
