+++ frontend/src/app/pages/split-payment/split-payment.component.scss
