+++ frontend/src/app/shared/components/basket-item/basket-item.component.html
@@ -1,0 +1,19 @@
+<div class="card">
+  <img [src]="item.image" />
+  <div class="information">
+    <div class="content">
+      <span class="title">{{ item.shortName }}</span>
+      <div class="sub-information">
+        <span class="quantity"
+          >{{ item.quantity }} {{ item.quantity > 1 ? 'items' : 'item' }}</span
+        >
+        <fa-icon
+          [icon]="trashIcon"
+          (click)="deleteFromBasket(item._id)"
+        ></fa-icon>
+      </div>
+    </div>
+
+    <span class="price">{{ item.price * item.quantity }}€</span>
+  </div>
+</div>
