+++ frontend/src/app/shared/components/basket/basket.component.scss
@@ -1,0 +1,38 @@
+.wrapper {
+  width: 100%;
+  height: 100%;
+  justify-content: center;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 1rem;
+  max-height: 60vh;
+}
+
+h2 {
+  margin: 0;
+}
+
+.items {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  overflow: auto;
+}
+
+.summary {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 1rem;
+
+  span {
+    font-weight: 800;
+  }
+}
+
+button {
+  width: 100%;
+}
