+++ frontend/src/app/shared/components/filter-allergens/filter-allergens.component.scss
