+++ frontend/src/app/shared/components/menu-item-detail/menu-item-detail.component.scss
@@ -1,0 +1,139 @@
+.item-image {
+  width: 100%;
+  height: 400px;
+  object-fit: cover;
+  display: block;
+  border: 1px solid #e0e0e0;
+  border-radius: 8px;
+}
+
+h2 {
+  text-align: left;
+  margin: 10px 0;
+  font-size: 1.6em;
+  font-weight: bold;
+}
+
+.ingredients-section {
+  margin: 15px 0;
+
+  h3 {
+    font-size: 1.3em;
+    margin-bottom: 15px;
+  }
+
+  ul {
+    list-style: none;
+    padding: 0;
+
+    li {
+      display: flex;
+      align-items: center;
+      margin-bottom: 15px;
+
+      app-counter {
+        flex-shrink: 0;
+      }
+
+      .ingredient-name {
+        font-size: 1.2em;
+        margin-left: 25px;
+        word-break: break-word;
+        flex-grow: 1;
+        min-width: 0;
+      }
+
+      .quantity-control {
+        display: flex;
+        align-items: center;
+
+        .remove-btn,
+        .add-btn {
+          color: white;
+          border: none;
+          border-radius: 4px;
+          width: 35px;
+          height: 35px;
+          cursor: pointer;
+          font-size: 2em;
+          display: flex;
+          align-items: center;
+          justify-content: center;
+        }
+
+        .remove-btn {
+          background-color: #c6230b;
+        }
+
+        .add-btn {
+          background-color: #00982b;
+        }
+
+        .quantity {
+          margin: 0 15px;
+          font-size: 1.5em;
+        }
+      }
+    }
+  }
+}
+
+.actions {
+  display: flex;
+  justify-content: flex-end;
+  align-items: center;
+  margin-top: 30px;
+  gap: 15px;
+
+  .delete-btn {
+    background-color: #c6230b;
+    color: white;
+    border: none;
+    border-radius: 8px;
+    width: 45px;
+    height: 45px;
+    cursor: pointer;
+    font-size: 1.1em;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+
+    fa-icon {
+      font-size: 1.5rem;
+    }
+  }
+
+  .delete-btn:disabled {
+    background-color: #ccc;
+    color: #666;
+    cursor: not-allowed;
+  }
+
+  .add-to-cart-btn {
+    background-color: #00982b;
+    color: white;
+    border: none;
+    border-radius: 8px;
+    height: 45px;
+    cursor: pointer;
+    font-size: 2rem;
+    padding: 0 25px;
+    font-weight: 500;
+  }
+}
+
+.quantity-selector select {
+  height: 45px;
+  appearance: none;
+  -webkit-appearance: none;
+  -moz-appearance: none;
+  padding: 0.4rem 1rem;
+  font-size: 1.5rem;
+  font-weight: bold;
+  color: rgb(76, 76, 76);
+  border-radius: 6px;
+  min-width: 60px;
+  transition:
+    border-color 0.2s,
+    box-shadow 0.2s;
+}
