+++ frontend/src/styles.scss
@@ -6,7 +6,10 @@
   --primary-color: #c6230b;
   --secondary-color: #ffffff;
   --base-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+
   --border-radius: 10px;
+  --base-border: 1px rgb(184, 184, 184) solid;
+  --footer-height: 10rem;
 }
 
 * {
